server/requirements.txt
flask
flask-cors
azure-ai-openai
